Fixed similary -> similarity key in Source
Well that was awkward to realize when I pulled in v2.0 in my own source
code. Apparently I typo'd in the Source type which affected the return
of the data from the lib as well. It was still being mapped to the API
result properly, but people implementing would have to have edited their
code without this, whoops?

Typos be gone (╯°□°）╯︵ ┻━┻

Signed-off-by: Jeroen Claassens <support@favware.tech>